Remove hardcoded `model.ast` classes in `ops` typeclasses to be able to use them with other languages. Addapts instances.
Rework `RefSyntax` to remove hardcoded reference to `SimplExpr`. To retain
the same usability as the old varargs version, add mutiple version of `apply`
with different numer of arguments (the `TupleInstances` are a hack to make
Scalac happy because was failing to find an implicit depending of the order
of the arguments ...)

Fix bug on `VarStatSyntax.relax`. If binary, must add respective constraints.

Add `StatRef` on `model.ast`.

Update dependencies.